Refactor Navbar Refresh Logic - Distinguish routes with and without Navbar - Update routes across the entire web application
