+++ client/src/Components/Login.js
@@ -63,7 +63,7 @@ function LandingPage(){
         });
         const data = await response.json();
         if(data.status==="forgot"){
-            navigate("/OTPverification", {state : email});
+            navigate("/verify/OTPverification", {state : email});
         }
         else{
             toast.warn("Email not found.");
@@ -107,7 +107,7 @@ function LandingPage(){
                             <button type="submit" className="btn">Login</button>
                         </div>
                     </div>
-                    <p className="not-register">Don't have an account? <Link to={"/Register"} className="links-color"> Register</Link> </p> 
+                    <p className="not-register">Don't have an account? <Link to={"/verify/Register"} className="links-color"> Register</Link> </p> 
                 </form>
             </div>
         </div>
