+++ client/src/Components/Profile.js
@@ -66,7 +66,7 @@ const Profile = () => {
 
   return (
     <>
-      <Navbar />
+      {/* <Navbar /> */}
       <div className="profile">
         <div className="profilePicture">
           {image ? (
